requirements/base.txt
django>=1.6.5,<1.7
south>=0.8.4,<0.9
pillow>=2.5.1,<2.6
django-stronghold>=0.2.6,<0.3
django-crispy-forms>=1.4.0,<1.5
django-authtools>=0.2.1,<0.3
django-tables2>=0.15.0,<0.16
django-braces>=1.4,<1.5
django-sekizai>=0.7,<0.8
django-codemirror-widget==0.4.0
fabric>=1.9,<1.10
logan>=0.5.10,<0.6
django-bootstrap-form>=3.1,<3.2
croniter==0.3.4
GitPython
gevent-socketio>=0.3.6,<0.4
virtualenv>=1.11.6,<1.12
python-social-auth>=0.1.26,<0.2.0
